Update service.ts adding callback in handlePong
If the sendPing function is called by the user, the callback function in there is never called if successful, only in terms of failure in the rpcFailure function (RPC failure, f.e. time out). So if the user calls sendPing the promise in the sendPing function never resolves as the callback is never called. The promise only can be rejected in terms of failure.

f.e. user calls sendPing expecting a promise, message, failure
const pong = await this.discv5.sendPing(enr)

By adding the check if the callback is set in the activeRequest in the handlePong function and calling the callback should solve this for the "success" case